feat: split database functions to separe module
